+++ docker/coturn/CHANGELOG.md
@@ -4,12 +4,63 @@ Coturn TURN server Docker image changelog
 
 
 
-## [4.5.2-r1] · 2021-04-?? · To-be-done
+## [4.5.2-r5] · 2021-08-29
+[4.5.2-r5]: /../../tree/docker/4.5.2-r5
+
+### Upgraded
+
+- [Debian Linux] "bullseye": <https://www.debian.org/releases/bullseye/releasenotes>
+
+
+
+
+## [4.5.2-r4] · 2021-08-28
+[4.5.2-r4]: /../../tree/docker/4.5.2-r4
+
+### Security updated
+
+- [Alpine Linux] 3.14.2: <https://github.com/docker-library/official-images/commit/b54effe0cc65795f29752ecc197328a04326a6f2>
+
+
+
+
+## [4.5.2-r3] · 2021-08-09
+[4.5.2-r3]: /../../tree/docker/4.5.2-r3
+
+### Security updated
+
+- [Alpine Linux] 3.14.1: <https://github.com/docker-library/official-images/commit/a52e64df36ca65954c6c7f57ad242634e834c73a>
+
+
+
+
+## [4.5.2-r2] · 2021-06-21
+[4.5.2-r2]: /../../tree/docker/4.5.2-r2
+
+### Upgraded
+
+- [Alpine Linux] 3.14: <https://alpinelinux.org/posts/Alpine-3.14.0-released.html>
+
+
+
+
+## [4.5.2-r1] · 2021-06-03
 [4.5.2-r1]: /../../tree/docker/4.5.2-r1
+
+### Added
+
+- [Prometheus] support with [prometheus-client-c] 0.1.3: <https://github.com/digitalocean/prometheus-client-c/releases/tag/v0.1.3> ([#754])
+
+### Improved
+
+- Use DNS requests to discover external IP address in `detect-external-ip` script ([#753]).
 
 ### Fixed
 
 - Incorrect linking with [mongo-c-driver] on [Debian Linux] image.
+
+[#753]: /../../pull/753
+[#754]: /../../pull/754
 
 
 
@@ -39,3 +90,5 @@ Coturn TURN server Docker image changelog
 [Coturn]: https://haraka.github.io
 [Debian Linux]: https://www.debian.org
 [mongo-c-driver]: https://github.com/mongodb/mongo-c-driver
+[Prometheus]: https://prometheus.io
+[prometheus-client-c]: https://github.com/digitalocean/prometheus-client-c